+++ server/src/klite/RequestLogger.kt
@@ -1,10 +1,16 @@
 package klite
 
+import java.lang.System.Logger.Level.ERROR
+import java.lang.System.Logger.Level.INFO
+
 typealias RequestLogFormatter = HttpExchange.(ms: Long) -> String?
 val defaultRequestLogFormatter: RequestLogFormatter = { ms ->
   "$remoteAddress $method $path$query: $statusCode in $ms ms - $browser" +
-    (failure?.takeIf { it !is StatusCodeException }?.let { " - $it" } ?: "")
+    (if (failure.isError) " - $failure" else "")
 }
+
+private val Throwable?.isError get() =
+  this != null && (this !is StatusCodeException || this.statusCode.isError && this !is NotFoundException)
 
 open class RequestLogger(
   val formatter: RequestLogFormatter = defaultRequestLogFormatter
@@ -15,7 +21,7 @@ open class RequestLogger(
     val start = System.nanoTime()
     exchange.onComplete {
       val ms = (System.nanoTime() - start) / 1000_000
-      formatter(exchange, ms)?.let { log.info(it) }
+      formatter(exchange, ms)?.let { log.log(if (exchange.failure.isError) ERROR else INFO, it) }
     }
     return handler(exchange)
   }